pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "armenian-voice-ai-bank-agent"
version = "0.1.0"
description = "Armenian Voice AI Support Agent for bank FAQs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.8.0",
  "pydantic-settings",
  "requests>=2.32.0",
  "beautifulsoup4>=4.12.0",
  "lxml>=5.2.0",
  "markdownify",
  "PyYAML",
  "httpx",
  "openai",
  "livekit-api",
  "livekit-agents",
  "livekit-plugins-openai",
]

[tool.setuptools]
packages = ["bank_support"]
py-modules = ["agent", "token_server"]
